% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{front}
\alias{front}
\title{plot gtclust pareto front}
\usage{
front(tree, mlogalpha = NULL)
}
\arguments{
\item{x}{a tree as produced by a gtclust variant.}
}
\value{
an \code{\link[ggplot2]{ggplot}} object
}
\description{
plot gtclust pareto front
}
