% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust-package.R
\docType{package}
\name{gtclust-package}
\alias{gtclust}
\alias{gtclust-package}
\title{gtclust: Hierarchal clustering of geographic or temporal object with contiguity constraints}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

gtclust offers two functions gtclust_poly and gtclust_graph that enable the clustering of spatial data such as polygons with a hclust type approach but taking advantages of contiguity constraints. The contiguities naturaly create a sparsely connected graph that can be leveraged to speed-up the calculations and deal with more than 30000 polygons in seconds.
}
\keyword{internal}
