% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\docType{package}
\name{gtclust}
\alias{gtclust}
\title{gtclust: A package for clustering spatial or temporal data with contiguity constrained hierarchical clustering}
\description{
The mypackage package provides three categories of important functions:
gtclust_graph, gtclust_temp, gt_poly.
}
\section{gtclust functions}{

The mypackage functions ...
}

