% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{gtclust_graph}
\alias{gtclust_graph}
\title{Hierarchical clustering with contiguity constraints}
\usage{
gtclust_graph(
  adjacencies_list,
  df,
  method = "ward",
  scaling = "raw",
  display_progress = FALSE
)
}
\arguments{
\item{adjacencies_list}{graph describing the contiguity between the rows of df as a list of adjacencies}

\item{df}{a data.frame with numeric columns}

\item{method}{linkage criterion in ward (default) or average, median}

\item{scaling}{default scaling of the features in zscore or raw (i.e. no scaling, the default)}

\item{display_progress}{boolean to set progression bar}
}
\value{
an \code{\link[stats]{hclust}} like object with additional slots
\describe{
  \item{data}{The numeric data (eventually scaled) used for the clustering}
  \item{centers}{The protoypes of each tree nodes}
}
}
\description{
This function take an data.frame and performs hierarchical clustering with contiguity 
constraints using a graph describing the contiguity (provided )
}
