% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{gtclust_poly}
\alias{gtclust_poly}
\title{Hierarchical clustering with contiguity constraints between polygons}
\usage{
gtclust_poly(
  df,
  method = "ward",
  adjacency = "rook",
  scaling = "raw",
  display_progress = FALSE
)
}
\arguments{
\item{df}{an \code{\link[sf]{sf}} data.frame with polygons like features}

\item{method}{linkage criterion in ward (default) or average, median}

\item{adjacency}{adjacency type to use  "rook" (default) or queen}

\item{scaling}{default scaling of the features in zscore (default) or raw (i.e. no scaling)}

\item{display_progress}{boolean to set progression bar}
}
\value{
an \code{\link[stats]{hclust}} like object with additional slots
\describe{
  \item{leafs_geometry}{geometries of the dendrogram leafs as an sfc list}
  \item{geotree}{geometries of the dendrogram no-leafs node as an sfc list}
  \item{data}{The numeric data (eventually scaled) used for the clustering}
  \item{centers}{The protoypes of each tree nodes}
}
}
\description{
This function take an \code{\link[sf]{sf}} data.frame and performs hierarchical clustering with contiguity constraints.
}
