% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{gtclust_poly_dcsbm}
\alias{gtclust_poly_dcsbm}
\title{Hierarchical clustering with contiguity constraints}
\usage{
gtclust_poly_dcsbm(
  poly.sf,
  ods,
  join_ori,
  join_dest,
  vol,
  lambda_in = sum(ods[vol])/(nrow(poly.sf)^2) * 100,
  lambda_ext = sum(ods[vol])/(nrow(poly.sf)^2)/100,
  adjacency = "rook",
  display_progress = FALSE
)
}
\arguments{
\item{display_progress}{boolean to set progression bar}

\item{adjacencies_list}{graph describing the contiguity between the rows of df as a list of adjacencies}

\item{triplet}{graph to fit with dcsbm in triplet fromat}
}
\value{
an \code{\link[stats]{hclust}} like object with additional slots
\describe{
  \item{data}{The numeric data (eventually scaled) used for the clustering}
  \item{centers}{The protoypes of each tree nodes}
}
}
\description{
This function take an data.frame and performs hierarchical clustering with contiguity 
constraints using a graph describing the contiguity (provided )
}
