% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{laplacian}
\alias{laplacian}
\title{Compute the laplacian of a graph}
\usage{
laplacian(A)
}
\arguments{
\item{A}{a sparse adjacency matrix}
}
\value{
A sparse matrix with the laplacian of A
}
\description{
This function compute the laplacian of a graph
}
