% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{log_nb_sptree}
\alias{log_nb_sptree}
\title{Compute the number of spanning tree in graph A of a graph}
\usage{
log_nb_sptree(A)
}
\arguments{
\item{A}{a sparse adjacency matrix}
}
\value{
the log of the number of spanning tree in graph A
}
\description{
This function compute the number of spanning tree in graph A
}
