% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{plot.gtclust}
\alias{plot.gtclust}
\title{plot gtclust dendrogram}
\usage{
\method{plot}{gtclust}(x, y = NULL, nb_leafs = 500, ...)
}
\arguments{
\item{x}{a tree as produced by gtclust. cutree() only expects a list with components merge, height, and labels, of appropriate content each.}

\item{nb_leaf}{number of leafs to keep}
}
\value{
an \code{\link[ggplot2]{ggplot}} object
}
\description{
plot gtclust dendrogram
}
