% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptree.R
\name{to_adjmat}
\alias{to_adjmat}
\title{Conversion from neighboring list to sparse adjacency matrix}
\usage{
to_adjmat(nb)
}
\arguments{
\item{nb}{neighboring list}
}
\value{
A sparse matrix
}
\description{
This function convert a neighboring list to sparse adjacency matrix
}
