% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{gtmethod_ward}
\alias{gtmethod_ward}
\alias{gtmethod_centroid}
\alias{gtmethod_median}
\alias{gtmethod_chisq}
\alias{gtmethod_bayes_mom}
\title{aggregation methods for gtclust}
\usage{
gtmethod_ward()

gtmethod_centroid()

gtmethod_median()

gtmethod_chisq()

gtmethod_bayes_mom(beta = 1)
}
\arguments{
\item{beta}{prior parameters for the dirichlet distribution}
}
\value{
An object of class gtmethod
}
\description{
aggregation methods for gtclust
}
\section{Functions}{
\itemize{
\item \code{gtmethod_ward}: classical ward method

\item \code{gtmethod_centroid}: classical centroid method

\item \code{gtmethod_median}: classical median method

\item \code{gtmethod_chisq}: chi-square method

\item \code{gtmethod_bayes_mom}: bayesian mixture of multinomials
}}

