% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{gtmethod_ward}
\alias{gtmethod_ward}
\alias{gtmethod_centroid}
\alias{gtmethod_median}
\alias{gtmethod_chisq}
\alias{gtmethod_bayes_mom}
\alias{gtmethod_bayes_dgmm}
\title{aggregation methods for gtclust}
\usage{
gtmethod_ward()

gtmethod_centroid()

gtmethod_median()

gtmethod_chisq()

gtmethod_bayes_mom(beta = 1)

gtmethod_bayes_dgmm(tau = 0.01, kappa = 1, beta = NaN, mu = as.matrix(NaN))
}
\arguments{
\item{beta}{Prior parameter (gamma rate), (default to NaN, in this case beta will be estimated from data as 0.1 time the mean of X columns variances)}

\item{tau}{Prior parameter (inverse variance), (default 0.01)}

\item{kappa}{Prior parameter (gamma shape), (default to 1)}

\item{mu}{Prior for the means (vector of size D), (default to NaN, in this case mu will be estimated from data as the mean of X)}
}
\value{
An object of class gtmethod
}
\description{
aggregation methods for gtclust
}
\section{Functions}{
\itemize{
\item \code{gtmethod_ward}: classical ward method

\item \code{gtmethod_centroid}: classical centroid method

\item \code{gtmethod_median}: classical median method

\item \code{gtmethod_chisq}: chi-square method

\item \code{gtmethod_bayes_mom}: bayesian mixture of multinomials

\item \code{gtmethod_bayes_dgmm}: bayesian diagonal gaussian mixture model
}}

