% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{gtmethod_ward}
\alias{gtmethod_ward}
\alias{gtmethod_centroid}
\alias{gtmethod_median}
\alias{gtmethod_chisq}
\alias{gtmethod_bayes_mom}
\title{ward method for gtclust}
\usage{
gtmethod_ward()

gtmethod_centroid()

gtmethod_median()

gtmethod_chisq()

gtmethod_bayes_mom(beta = 1)
}
\arguments{
\item{beta}{prior parameters for the dirichlet}
}
\value{
An object of class gtmethod

An object of class gtmethod

An object of class gtmethod

An object of class gtmethod

An object of class gtmethod
}
\description{
ward method for gtclust

ward method for gtclust

ward method for gtclust

ward method for gtclust

ward method for gtclust
}
\section{Functions}{
\itemize{
\item \code{gtmethod_ward}: classical ward method

\item \code{gtmethod_centroid}: classical entroid method

\item \code{gtmethod_median}: classical median method

\item \code{gtmethod_chisq}: chisq method

\item \code{gtmethod_bayes_mom}: bayesian mixture of multinomial
}}

