% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtclust.R
\name{plot.gtclust}
\alias{plot.gtclust}
\title{plot gtclust dendrogram}
\usage{
\method{plot}{gtclust}(x, y = NULL, nb_max_leafs = 500, ...)
}
\arguments{
\item{x}{a tree as produced by a gtclust variant.}

\item{nb_max_leafs}{number of leafs to keep}
}
\value{
an \code{\link[ggplot2]{ggplot}} object
}
\description{
plot gtclust dendrogram
}
